Fix more codacy issues
