+++ boggs/src/main/java/com/hartwig/healthchecks/boggs/healthcheck/prestasts/PrestatsHealthChecker.java
@@ -1,14 +1,11 @@
 package com.hartwig.healthchecks.boggs.healthcheck.prestasts;
 
 import java.io.IOException;
-import java.util.List;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.jetbrains.annotations.NotNull;
 
-import com.hartwig.healthchecks.boggs.model.report.PrestatsDataReport;
-import com.hartwig.healthchecks.boggs.model.report.PrestatsReport;
 import com.hartwig.healthchecks.common.checks.HealthChecker;
 import com.hartwig.healthchecks.common.exception.EmptyFileException;
 import com.hartwig.healthchecks.common.util.BaseReport;
@@ -18,10 +15,6 @@ import com.hartwig.healthchecks.common.util.ErrorReport;
 public class PrestatsHealthChecker implements HealthChecker {
 
     protected static final String ERROR_MSG = "Got An Exception with message: %s";
-
-    private static final String FAIL_ERROR = "FAIL";
-
-    private static final String FOUND_FAILS_MSG = "NOT OK: %s has status FAIL for Patient %s";
 
     private static final Logger LOGGER = LogManager.getLogger(PrestatsHealthChecker.class);
 
@@ -39,23 +32,14 @@ public class PrestatsHealthChecker implements HealthChecker {
     @Override
     public BaseReport runCheck() {
 
-        PrestatsReport prestatsReport;
+        BaseReport prestatsReport;
         try {
             prestatsReport = dataExtractor.extractFromRunDirectory(runDirectory);
         } catch (IOException | EmptyFileException exception) {
             LOGGER.error(String.format(ERROR_MSG, exception.getMessage()));
-            return new ErrorReport(CheckType.PRESTATS, exception.getClass().getName(), exception.getMessage());
+            prestatsReport = new ErrorReport(CheckType.PRESTATS, exception.getClass().getName(),
+                            exception.getMessage());
         }
-        logPrestatsReport(prestatsReport.getReferenceSample());
-        logPrestatsReport(prestatsReport.getTumorSample());
         return prestatsReport;
-    }
-
-    private void logPrestatsReport(final List<PrestatsDataReport> prestatsDataReport) {
-        prestatsDataReport.forEach((prestatsData) -> {
-            if (prestatsData.getStatus().equalsIgnoreCase(FAIL_ERROR)) {
-                LOGGER.info(String.format(FOUND_FAILS_MSG, prestatsData.getCheckName(), prestatsData.getPatientId()));
-            }
-        });
     }
 }
